Squashed commit of the following:
commit 726ab3720feb8e189a13cff8c561bb2ab0c14558
Author: alex <darkstar79@gmx.net>
Date:   Wed Mar 18 23:03:52 2026 +0100

    Update TRAINING_MODE.md to reference SE difficulty rating scale


commit e1a98b2fdf9df7895b75cacf2edf5d9df2c7d3a3
Author: alex <darkstar79@gmx.net>
Date:   Wed Mar 18 22:16:51 2026 +0100

    Adopt Sudoku Explainer (SE) difficulty rating standard

    Replace custom sum-based scoring with the established SE rating system
    where puzzle difficulty equals the hardest technique required to solve
    it (max, not sum). Uses authentic SE v1.2.1 ratings on the 1.0-12.0
    scale by Nicolas Juillerat, with SE-compatible ratings for techniques
    not in the original SE.

+++ src/core/statistics_manager.cpp
@@ -81,7 +81,7 @@ StatisticsManager::~StatisticsManager() {
 }
 
 std::expected<uint64_t, StatisticsError> StatisticsManager::startGame(Difficulty difficulty, uint32_t puzzle_seed,
-                                                                      int puzzle_rating) {
+                                                                      double puzzle_rating) {
     if (!isValidDifficulty(difficulty)) {
         return std::unexpected(StatisticsError::InvalidDifficulty);
     }
@@ -504,7 +504,7 @@ void StatisticsManager::updateAggregateStats(const GameStats& completed_game) co
     int diff_index = static_cast<int>(completed_game.difficulty);
 
     // Update rating statistics BEFORE incrementing games_played (track for all games, not just completed)
-    if (completed_game.puzzle_rating > 0) {
+    if (completed_game.puzzle_rating > 0.0) {
         // Get count BEFORE incrementing
         auto games_count_before = cached_stats_.games_played[diff_index];
